Additional try to locate monolog package
Having configured the composer.json of my app with

      "config": {"vendor-dir": "Vendor"}

which I assume is a fairly common option, the monolog package is located in app/Vendor. And the plugin cannot find it.

Of course with the 'search' param in CakeLog::config this can be solved, but maybe is better to take care of this directly in the plugin.